+++ changelog.md
@@ -1,10 +1,10 @@
 ### v0.3.0
-
 - Improved status notification by adding progress bar and detailed info.
+- Refactored commands system, improving maintainability.
 - Added fuzzy search when entering plugin commands
+- Added context filtered results, e.g. 'po start' won't be an shown on an already running session.
 
 ### v0.2.0
-
 - Major refactor of the plugin, improving maintainability.
 - Added pausing of a session.
 - Added displaying status of current session.