+++ changelog.md
@@ -1,7 +1,7 @@
 ### v0.3.0
 - Improved status notification by adding progress bar and detailed info.
 - Refactored commands system, improving maintainability.
-- Added fuzzy search when entering plugin commands
+- Added fuzzy search when typing plugin commands
 - Added context filtered results, e.g. 'po start' won't be an shown on an already running session.
 
 ### v0.2.0